Add my optimized copy of screen/neofetch, called nanofetch
nanofetch is made to be simple, and fitted to my needs.
I don't own the copyrights on the logos, but the colors are by me.

screenfetch takes almost 500 ms to display the same informations,
which was too long for me. So this tool does the same, but in less
than 100 ms (a bit more when either glxinfo or lspci are slow. I
don't know the reason, it seems to be random).

Expect further optimizations to reduce execution time!